fix: refactor reticle_gen
