+++ packages/reticle_gen/lib/mil_xt_unoptimized.dart
@@ -1,4 +1,3 @@
-import 'package:reticle_gen/mil_reticle.dart';
 import 'package:reticle_gen/reticle_gen.dart';
 
 const double epsilon = 1e-6;
@@ -109,13 +108,13 @@ class MilXtVariant {
   }
 }
 
-class MilXtReticleDrawer extends MilReticleDrawer {
+class MilXtReticleDrawer implements SVGDrawerInterface {
   final MilXtVariant variant;
 
   MilXtReticleDrawer({this.variant = MilXtVariant.defaultVariant});
 
   @override
-  void draw(CanvasInterface canvas) {
+  void draw(SVGCanvas canvas) {
     final A = variant.a;
     final F = variant.f;
     final K = variant.k;
@@ -399,7 +398,7 @@ void main(List<String> args) {
   print('Generating "${variant.name}"  →  $outputPath');
   print('  A=${variant.a}  F=${variant.f}  K=${variant.k}');
 
-  MilReticleCanvas(milWidth: 48, milHeight: 48)
+  MilReticleSVGCanvas(milWidth: 48, milHeight: 48)
     ..generate(MilXtReticleDrawer(variant: variant))
     ..svg.export(outputPath);
 }
